aif2pcm: Fix length reporting. Don't use a MARK chunk if no loop.
The reported sample length in the aif header can differ from the actual
length of the SSND chunk. The END marker was being misused to contain
the reported length, but this still counts as a loop in aif2agb, even
if the START marker is omitted (it defaults to the END value).

Now the reported length is in the COMM chunk, which is more correct.